[CL-3928] Fixed generate command and moved formatter to separate service
