+++ back/engines/free/email_campaigns/app/models/email_campaigns/campaigns/voting_phase_started.rb
@@ -71,13 +71,10 @@ module EmailCampaigns
       if notification.phase.voting?
         [{
           event_payload: {
+            project_url: Frontend::UrlService.new.model_to_url(notification.phase.project, locale: recipient.locale),
+            project_title_multiloc: notification.phase.project.title_multiloc,
             phase_title_multiloc: notification.phase.title_multiloc,
-            phase_description_multiloc: notification.phase.description_multiloc,
-            phase_start_at: notification.phase.start_at.iso8601,
-            phase_end_at: notification.phase.end_at.iso8601,
-            phase_url: Frontend::UrlService.new.model_to_url(notification.phase, locale: recipient.locale),
-            project_title_multiloc: notification.project.title_multiloc,
-            project_description_preview_multiloc: notification.project.description_preview_multiloc
+            ideas: EmailCampaigns::PayloadFormatterService.new.format_ideas_list(notification.phase.ideas, recipient)
           },
           delay: 8.hours.to_i
         }]
